+++ cookbooks/aws-parallelcluster-awsbatch/metadata.rb
@@ -9,10 +9,6 @@ source_url 'https://github.com/aws/aws-parallelcluster-cookbook'
 chef_version '>= 18'
 version '3.10.0'
 
-supports 'amazon', '>= 2.0'
-supports 'centos', '>= 7.0'
-supports 'ubuntu', '>= 20.04'
-
 depends 'apt', '~> 7.5.22'
 depends 'iptables', '~> 8.0.0'
 depends 'nfs', '~> 5.1.2'