+++ cookbooks/aws-parallelcluster-environment/recipes/config/network_interfaces.rb
@@ -12,7 +12,8 @@
 # OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
 
-return if on_docker?
+# amazon-ec2-net-utils is pre-installed in AL2023 and handles multi-nics instances properly
+return if on_docker? || platform?('amazon') && node['platform_version'].to_i == 2023
 
 def network_card_index(mac, token)
   uri = URI("http://169.254.169.254/latest/meta-data/network/interfaces/macs/#{mac}/network-card")