+++ cookbooks/aws-parallelcluster-platform/resources/nvidia_driver/nvidia_driver_amazon2.rb
@@ -22,5 +22,13 @@ def set_compiler?
 end
 
 def compiler_version
+  'gcc10'
+end
+
+def compiler_path
   'CC=/usr/bin/gcc10-gcc'
+end
+
+def extra_packages
+  %w()
 end