+++ CHANGELOG.md
@@ -8,7 +8,6 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 
 **ENHANCEMENTS**
 - Add support for external Slurmdbd.
-- Add support for FSx Lustre in US isolated regions.
 
 **CHANGES**
 - Upgrade Cinc Client to version to 18.4.12 from 18.2.7.
@@ -35,6 +34,27 @@ This file is used to list changes made in each version of the AWS ParallelCluste
   EFS is used for shared internal data. 
 - On Ubuntu systems, remove default logrotate configuration for cloud-init log files that clashed with the
   configuration coming from Parallelcluster.
+- Removing `/etc/profile.d/pcluster.sh` so that it's not executed at every user login and 
+  `cfn_bootstrap_virtualenv` is not added in PATH environment variable.
+
+3.9.3
+------
+
+**ENHANCEMENTS**
+- Add support for FSx Lustre as a shared storage type in us-iso-east-1.
+
+**BUG FIXES**
+- Remove `cloud_dns` from the `SlurmctldParameters` in the Slurm config to avoid Slurm fanout issues.  
+  This is also not required since we set the IP addresses on instance launch.
+
+3.9.2
+------
+
+**CHANGES**
+- Upgrade Slurm to 23.11.7 (from 23.11.4).
+
+**BUG FIXES**
+- Fix issue with `SharedStorageType: Efs` not bootstrapping on ARM instances.
 
 3.9.1
 ------